Change SelectionPercentage from double to int
Refactor the data type of SelectionPercentage to an integer for consistency across the DTO and service code. This change ensures type compatibility and avoids potential precision errors when handling percentage values.

+++ DistLearning.Service/TestsService.cs
@@ -269,7 +269,7 @@ public class TestsService(ITestRepository testRepository, IMapper mapper) : ITes
                 Uuid = answerOption.Uuid,
                 OptionText = answerOption.OptionText,
                 SelectionCount = selectionCount,
-                SelectionPercentage = percentage,
+                SelectionPercentage = (int)percentage,
                 IsAllowedText = answerOption.IsAllowedText,
             };
